+++ lib/topics/plans/controller.dart
@@ -11,13 +11,13 @@ class PlansController {
   Future<List<StudyTeamEntry>> list() async {
     var url = student.buildUrl(
         'studieplan.aspx?displaytype=ugeteksttabel&elevid=${student.studentId}');
-    var req = await lppDio.get(url);
+    var req = await request(url);
     return extractPlans(BeautifulSoup(req.data));
   }
 
   Future<StudyPlanEntry> get(StudyPlanRef ref) async {
     var url = student.buildUrl('studieplan/forloeb_vis.aspx?phaseid=${ref.id}');
-    var req = await lppDio.get(url);
+    var req = await request(url);
     return extractPlanEntry(BeautifulSoup(req.data));
   }
 }