+++ lib/topics/teams/controller.dart
@@ -10,14 +10,14 @@ class TeamsController {
 
   Future<List<Team>> list() async {
     var url = student.buildUrl('forside.aspx');
-    var req = await lppDio.get(url);
+    var req = await request(url);
     return extractTeams(BeautifulSoup(req.data));
   }
 
   Future<ModuleStatistics?> get(Team team) async {
     var url =
         student.buildUrl('subnav/modulregnskab.aspx?holdelementid=${team.id}');
-    var req = await lppDio.get(url);
+    var req = await request(url);
     return extractModuleStats(BeautifulSoup(req.data));
   }
 }