+++ test/lectio_wrapper_test.dart
@@ -1,5 +1,3 @@
-import 'dart:io';
-
 import 'package:dotenv/dotenv.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:lectio_wrapper/lectio_wrapper.dart';
@@ -41,16 +39,10 @@ void main() {
         gyms.firstWhere((element) => element.name == "Egaa Gymnasium").id, 256);
   });
 
-  test('getFile()', () async {
-    var file = await student!.getFile(
-        "https://www.lectio.dk/lectio/256/lc/60231186938/res/60231186939/RetteGuide%20Afleveringer%201g.docx%202.docx");
-    File lFile = File("/Users/oscar/dev/lectio_wrapper/out/file.docx");
-    await lFile.writeAsBytes(file);
-    expect(file, isNotEmpty);
-  });
-
   test('getCookies()', () async {
     var cookies = await student!.getCookies();
+    var cookieString =
+        cookies.map((cookie) => "${cookie.name}=${cookie.value}").join(',');
     expect(cookies, isNotNull);
   });
 }